Chore: Moved API key models to `permissions.py`
