+++ pydatalab/src/pydatalab/permissions.py
@@ -1,3 +1,4 @@
+import datetime
 from functools import wraps
 from hashlib import sha512
 from typing import Any
@@ -11,9 +12,52 @@ from pydatalab.config import CONFIG
 from pydatalab.logger import LOGGER
 from pydatalab.login import UserRole
 from pydatalab.models.people import AccountStatus
+from pydatalab.models.utils import BaseModel, PyObjectId
 from pydatalab.mongo import flask_mongo, get_database
 
 PUBLIC_USER_ID = ObjectId(24 * "0")
+
+
+class Key(BaseModel):
+    user: PyObjectId
+    """id that the object belongs to"""
+
+    created_at: "datetime.datetime"
+    """date that the object was created"""
+
+    expires_at: "datetime.datetime|None"
+    """date that the object expires"""
+
+    version: int
+    """version of the key"""
+
+
+class AccessToken(Key):
+    refcode: str
+    """The reference code of the access key"""
+
+    active: bool
+    """whether the access key is active"""
+
+    type: str = "access_token"
+    """the type of access key"""
+
+    token: str
+    """The hash of the access token"""
+
+
+class ApiKey(Key):
+    name: str
+    """The name of the API key"""
+
+    digest: str
+    """The hash of the API key"""
+
+    type: str = "api_key"
+    """The type of the API key"""
+
+    hash: str
+    """The hash of the token"""
 
 
 def active_users_or_get_only(func):
